Move home components to home directory Add auth user name in header
